todo_app-2/requirements.txt
fastapi
uvicorn[standard]
pika
SQLAlchemy
python-dotenv
opentracing
jaeger-client
starlette>=0.27.0
opentracing-instrumentation
strawberry-graphql[fastapi]>=0.200.0
strawberry-graphql[dataloader]
strawberry-graphql[federation]
asyncio